Code to try get Event working

+++ TedXEventApp/src/main/java/com/tedx/capetown/app/core/models/EventModel.java
@@ -11,8 +11,8 @@ public class EventModel
     public String imageURL;
     public String websiteURL;
     public String descriptionHTML;
-    public long latitude;
-    public long longitude;
+    public double latitude;
+    public double longitude;
     public String locationDescriptionHTML;
     public SessionCollectionModel sessions;
 }
